requirements.txt
Django==4.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.8.6
python-dotenv==1.0.0